Add Chocolate Fluid, Fix Fast Leaf Decay
Added leaf block tags.
Not experienced enough to figure out why the texture on the logs isn't rotating. They're registered as RotatedPillarBlock so they should work. Could be a model issue instead but I remember having problems even getting the model textures to work in the first place so leaving for now.