remove instant health
Giving the effect instant health to a consumable is unbalanced, the player can essentially spam eating to stay alive indefinitely. in addition, it also makes potions of instant health obsolete. 

I made this PR to suggest removing instant health from Jaffa Cakes as regen and health boost is strong enough as is.

Cheers!